+++ Data/EventData/00_Mainline/11_TheWarWithin.lua
@@ -4,18 +4,10 @@ local type = addon.Objects.EventType;
 
 KrowiAF.EventData.TheWarWithin = {
     { -- WoW's 20th Anniversary
-        1500, type.Calendar,
-        1173,
-        133783,
+        1595, type.Calendar,
+        1678,
+        6036545,
 		addon.L["WoW's 20th Anniversary"],
-        addon.L["Other"],
-        2
-    },
-    { -- WoW's 21st Anniversary
-        1501, type.Calendar,
-        1173,
-        133783,
-		addon.L["WoW's 21st Anniversary"],
         addon.L["Other"],
         2
     },